[Partitioner] Partitioner Refactoring 1
