+++ lib/Partitioner/CMakeLists.txt
@@ -1,5 +1,7 @@
 add_library(Partitioner
               PartitionerUtils.cpp
+              PartitionerOptimizer.cpp
+              PartitionerValidation.cpp
               Partitioner.cpp)
 
 target_link_libraries(Partitioner
